+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Avoid including certconfig UID in organizations for kubeconfig requests.
+
 ## [2.0.1] - 2022-04-04
 
 ### Fixed